requirements/dev.txt
-r common.txt

black==22.10.0
click==8.1.3
mypy==0.991
mypy-extensions==0.4.3
pycodestyle==2.9.1
pyflakes==2.5.0
types-PyYAML==6.0.12.2
typing==3.7.4.3
typing_extensions==4.4.0
django-stubs==1.13.0
django-stubs-ext==0.7.0
flake8==5.0.4
django-shortcuts==1.6
types-pytz==2022.6.0.1
python-dotenv==0.21.0
djlint==1.19.7
pyparsing==3.0.9
pydot==1.4.2
factory-boy==3.2.1